requirements.txt
pandas
PyYAML
ttkbootstrap